Add support for `savePath` in Cloudinary CDN and prompt to replace file
Updated the Cloudinary CDN uploader to use the `savePath` argument when uploading a file. This may cause conflicts with existing files that have already been uploaded. If the file already exists, the extension will prompt the user to confirm if they want to overwrite the existing file. If yes, then the file will be re-uploaded with the option `overwrite: true`.