Changed configuration layout to provide new UI
Since there are many settings to configure so it is really user friendly
if we can show the sections in extension navigation. This change is
using different configuration layout in package.json to support the
tabbed UI. getConfig function also needed a change to get the
configurations from new layout.